requirements.txt
Flask==3.1.0
flask-pymongo==2.3.0
python-dotenv==1.0.0